Minnor changes

+++ onboard/src/xcs/nodes/red_dot.xob/x_red_dot.hpp
@@ -24,8 +24,8 @@ namespace reddot{
 
         XVar<::urbi::UImage> enhancedVideo;
         XVar<bool> found;
-        XVar<int> errorX;
-        XVar<int> errorY;
+        XVar<double> errorX;
+        XVar<double> errorY;
 
         XRedDot(const std::string &name);
     };
